Extend protobuf schema
Include a field in lighthouse announce that specifies
the public mapped address.

On startup configure this address mapping and let the
portmapper ensure it is up to date on its own.

Every lighthouse announce interval check if the current
network supports NAT-PMP based on portmapper's attempts
to map the port, and if it is then send the mapped
port and public IP address along in the announce
request.